requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pinecone-client==4.1.0
openai==1.51.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3